% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar-notebook.R
\name{notebook_browse}
\alias{notebook_browse}
\title{Open the notebook in a browser}
\usage{
notebook_browse(file = NULL)
}
\arguments{
\item{file}{full path to the notebook file. Defaults to \code{NULL} in which case
it finds the path using targets.}
}
\value{
This function is called for its side effects so it return \code{NULL}
invisibly.
}
\description{
Open the notebook in a browser
}
