% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-notestar.R
\name{notebook_create_page}
\alias{notebook_create_page}
\title{Create a new notebook page}
\usage{
notebook_create_page(slug = NULL, date = NULL, open = TRUE)
}
\arguments{
\item{slug}{Optional "slug" (label) for the post. Defaults to \code{""}.}

\item{date}{Optional data to use. This date should have the format
\code{YYYY-MM-DD}. Defaults to the current date.}

\item{open}{Whether to open the new file for editing. Defaults to
\code{rlang::is_interactive()}.}
}
\value{
Invisible returns the relative path to the created file.
}
\description{
Creates a file with pattern \verb{[notebook_dir]/[date][-slug].Rmd}.
}
