% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar-notebook.R
\name{notebook_knit_page}
\alias{notebook_knit_page}
\title{Knit a single notebook entry}
\usage{
notebook_knit_page(rmd_in, md_out, helper_script)
}
\arguments{
\item{rmd_in}{path to the rmd file to knit}

\item{md_out}{path to the md file to create}

\item{helper_script}{path to an R script to run beforehand}
}
\description{
Knit a single notebook entry
}
\keyword{internal}
