% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar-notebook.R
\name{notebook_write_yaml}
\alias{notebook_write_yaml}
\title{Write out a yaml-file and return the path}
\usage{
notebook_write_yaml(x, file, ...)
}
\arguments{
\item{x, file, ...}{arguments passed to \code{yaml::write_yaml()}.}
}
\value{
the yaml \code{file} is written and the value of \code{file} is returned
}
\description{
This function is a wrapper over \code{yaml::write_yaml()} that works with file
targets.
}
