% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notestar-package.R
\docType{package}
\name{notestar-package}
\alias{notestar}
\alias{notestar-package}
\title{notestar: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjmahr/notestar}
  \item Report bugs at \url{https://github.com/tjmahr/notestar/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
