% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr-helpers.R
\name{notebook_set_opts_knit}
\alias{notebook_set_opts_knit}
\alias{notebook_set_opts_chunk}
\alias{notebook_set_markdown_hooks}
\title{Set default knitr options for notebook knitting}
\usage{
notebook_set_opts_knit()

notebook_set_opts_chunk()

notebook_set_markdown_hooks()
}
\value{
a list containing the default options used
}
\description{
These options are used in the \code{knitr-helpers.R} file.
}
