% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_notebook.R
\name{tar_notebook}
\alias{tar_notebook}
\title{Assemble knitted notebook md files into a single-page bookdown document}
\usage{
tar_notebook(
  dir_notebook = "notebook",
  dir_md = "notebook/book",
  theme = "water",
  book_filename = "notebook",
  subdir_output = "docs",
  extra_deps = list()
)
}
\arguments{
\item{dir_notebook}{Name of the directory containing the Rmd files. It should
be a relative path from the project root. Defaults to \code{"notebook"}}

\item{dir_md}{Name of the directory to contain md files (knitted Rmd files).
It should be a relative path from the project root. Defaults to
\code{"notebook/book"}.}

\item{theme}{Theme to use for \code{cleanrmd::html_document_clean()}. Defaults to
\code{"water"}.}

\item{book_filename}{Name to use for the final html file. Defaults to
\code{"notebook"} which produces \code{"notebook.html"}.}

\item{subdir_output}{Subdirectory of \code{dir_md} which will contain final html
file produced by bookdown. Defaults to \code{"docs"}.}

\item{extra_deps}{A list of extra dependencies. These should be the names of
targets defined elsewhere in the dependencies graph. Defaults to \code{list()}.
Use this argument, for example, to force a notebook to depend on a \code{.bib}
file.}
}
\value{
A list of targets.
}
\description{
Assemble knitted notebook md files into a single-page bookdown document
}
\details{
The list of targets produced includes:
\itemize{
\item \code{notebook_output_yaml}, a file target for \verb{_output.yml} (used by
bookdown to set output format options).
\item \code{notebook_bookdown_yaml}, a file target for \verb{_bookdown.yml} (used by
bookdown to collate pages).
\item \code{notebook}, a file target for the final assembled html file used by
bookdown.
}

The only output format supported is an html file produced by
\code{cleanrmd::html_document_clean()}.
}
