% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar-notebook.R
\name{tar_notebook_index_rmd}
\alias{tar_notebook_index_rmd}
\title{Create index.Rmd file}
\usage{
tar_notebook_index_rmd(
  title = "Notebook title",
  author = "Author name",
  bibliography = ymlthis::yml_blank(),
  csl = ymlthis::yml_blank(),
  ...,
  .list = rlang::list2(...),
  dir_notebook = NULL,
  dir_md = NULL
)
}
\arguments{
\item{title}{Title to use for the notebook. Defaults to \code{"Notebook title"}.}

\item{author}{Author name for the notebook. Defaults to \code{"Author Name"}.
Multiple authors, email and affiliations can used by using a list \code{list(name = names, email = emails)}. See \code{\link[ymlthis:yml_author]{ymlthis::yml_author()}}.}

\item{bibliography, csl}{Name of a \code{.bib} file in the \code{dir_notebook} rmd folder.
This file is copied to \verb{[dir_md]/assets/[bibliography]} and saved in the
YAML metadata in index.Rmd as "./assets/\link{bibliography}". Defaults to
\code{\link[ymlthis:yml_blank]{ymlthis::yml_blank()}}. This file is then tracked with
targets and set as a dependency for \code{tar_notebook()}.}

\item{csl}{Name of a \code{.csl} file in the \code{dir_notebook} rmd folder.
This file is copied to \verb{[dir_md]/assets/[csl]} and saved in the
YAML metadata in index.Rmd as "./assets/\link{csl}". Defaults to
\code{\link[ymlthis:yml_blank]{ymlthis::yml_blank()}}. This file is then tracked with
targets and set as a dependency for \code{tar_notebook()}.}

\item{...}{Additional key-value pairs for setting fields in the yml metadata.
For example, \code{subtitle = "Status: Project completed"} would add a subtitle
to the index.Rmd file and that subtitle would appear in the notebook. These
fields cannot include \code{"date"} or \code{"site"}.}

\item{.list}{Alternatively, these YAML fields may be set using a list. Any
key-value pairs in this list will override options set by arguments. For
example, an \code{author} entry in \code{.list} would override the value in the
\code{author} parameter.}

\item{dir_notebook}{The notebook directory. Defaults to \code{"notebook"}.}
}
\value{
a list of targets:
\itemize{
\item \code{notebook_bibliography_user}, \code{notebook_csl_user}: file targets for
the bibliography and csl files in the \code{dir_notebook} user folder.
\item \code{notebook_bibliography_asset}, \code{notebook_csl_asset}: file targets for
the bibliography and csl files in the \code{dir_md} assets folder. These
command to build these targets to copy them.
\item \code{notebook_deps_in_index_yml}: a list with the names of asset
bibliography and asset csl targets if they are present. Can be an
empty list.
\item \code{notebook_index_rmd}: file target for the index.Rmd file.
}

If any of these file targets is not used, an empty target \code{list()} is passed
along.
}
\description{
This function creates the \code{index.Rmd} file for the notebook and sets YAML
metadata in the file. It also detects CSL and bibliography file dependencies.
}
