% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-notestar.R
\name{use_notestar}
\alias{use_notestar}
\alias{use_notestar_makefile}
\title{Create a brand new notebook with default settings}
\usage{
use_notestar(
  dir_project = ".",
  dir_notebook = "notebook",
  dir_md = "notebook/book",
  notebook_helper = "knitr-helpers.R",
  cleanrmd_theme = "water",
  notebook_filename = "notebook",
  open = interactive()
)

use_notestar_makefile(dir_project = ".")
}
\arguments{
\item{dir_project}{file-path to the base/root folder of the project. Defaults to
\code{"."} which is the current working directory.}

\item{dir_notebook}{Name of the directory containing the Rmd files. It should
be a relative path from the project root. Defaults to \code{"notebook"}}

\item{dir_md}{Name of the directory to contain md files (knitted Rmd files).
It should be a relative path from the project root. Defaults to
\code{"notebook/book"}.}

\item{notebook_helper}{Filename for an R script to run before knitting each
Rmd file and rendering the notebook with bookdown. The file must be in
\code{dir_md}. Defaults to \code{"knitr-helpers.R"} so the default location is
\verb{"notebook/knitr_helpers.R}.}

\item{cleanrmd_theme}{Theme to use for \code{cleanrmd::html_document_clean()}. Defaults to
\code{"water"}.}

\item{notebook_filename}{Name to use for the final html file. Defaults to
\code{"notebook"} which produces \code{"notebook.html"}.}

\item{open}{whether to open \verb{_targets.R}, \code{notebook/index.Rmd}, and
\code{R/functions.R} when they are created. Defaults to \code{interactive()} (whether
the code is being called interactively).}
}
\value{
\code{use_notestar_makefile()} creates a Makefile that will build or clean a
targets-based workflow.
}
\description{
Create a brand new notebook with default settings
}
