% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar-notebook.R
\name{notebook_rmd_collate}
\alias{notebook_rmd_collate}
\title{Gather rmd entries and put them in order}
\usage{
notebook_rmd_collate(dir_notebook = "notebook")
}
\value{
a vector of paths to rmd files
}
\description{
Files are sorted with index.Rmd first followed by the rest in lexicographic
order.
}
\keyword{internal}
