% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notestar-package.R
\docType{package}
\name{notestar-package}
\alias{notestar}
\alias{notestar-package}
\title{notestar: Notebooks Using 'Targets' and 'Bookdown'}
\description{
'Targets' is an R package for dependency and build management in 
    data analysis projects. This package provides a set of targets and project 
    infrastructure to create 'bookdown'-based notebooks using 'targets'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjmahr/notestar}
  \item Report bugs at \url{https://github.com/tjmahr/notestar/issues}
}

}
\author{
\strong{Maintainer}: Tristan Mahr \email{tristan.mahr@wisc.edu} (\href{https://orcid.org/0000-0002-8890-5116}{ORCID})

}
\keyword{internal}
