% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar-notebook.R
\name{tar_notebook}
\alias{tar_notebook}
\title{Assemble knitted notebook md files into a single-page bookdown document}
\usage{
tar_notebook(
  subdir_output = "docs",
  extra_deps = list(),
  use_downlit = FALSE,
  markdown_document2_args = list()
)
}
\arguments{
\item{subdir_output}{Subdirectory of \code{dir_md} which will contain final html
file produced by bookdown. Defaults to \code{"docs"}.}

\item{extra_deps}{A list of extra dependencies. These should be the names of
targets defined elsewhere in the dependencies graph. Defaults to \code{list()}.
Use this argument, for example, to force a notebook to depend on a \code{.css}
file.}

\item{use_downlit}{whether to post-process the notebook with the downlit
syntax highlighter. Default is \code{FALSE}.}

\item{markdown_document2_args}{arguments to pass onto
\code{\link[bookdown:html_document2]{bookdown::markdown_document2()}}. Defaults to \code{list()}.}
}
\value{
A list of targets.
}
\description{
Assemble knitted notebook md files into a single-page bookdown document
}
\details{
The list of targets produced includes:
\itemize{
\item \code{notebook_output_yaml}, a file target for \verb{_output.yml} (used by
bookdown to set output format options).
\item \code{notebook_bookdown_yaml}, a file target for \verb{_bookdown.yml} (used by
bookdown to collate pages).
\item \code{notebook}, a file target for the final assembled html file used by
bookdown.
}

The only output format supported is an html file produced by
\code{cleanrmd::html_document_clean()}.
}
