% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar-notebook.R
\name{tar_notebook_pages}
\alias{tar_notebook_pages}
\title{Create targets to knit notebook Rmd files}
\usage{
tar_notebook_pages(dir_notebook = NULL, dir_md = NULL, notebook_helper = NULL)
}
\value{
A list of targets.
}
\description{
Create targets to knit notebook Rmd files
}
\details{
The list of targets produced includes:
\itemize{
\item \code{notebook_helper}, a file for the helper R script.
\item one file target for each input Rmd file. Any targets used
with \code{tar_read()} or \code{tar_load()} inside these files are detected
and checked for changes.
\item one file target for each output md file.
\item \code{notebook_rmds}, a combined target for the input Rmd files.
\item \code{notebook_pages}, a combined target for the output md files.
}
}
