% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-notestar.R
\name{use_notestar}
\alias{use_notestar}
\alias{use_notestar_makefile}
\title{Create a brand new notebook with default settings}
\usage{
use_notestar(
  dir_project = ".",
  dir_notebook = "notebook",
  dir_md = "notebook/book",
  notebook_helper = "notebook/book/knitr-helpers.R",
  open = interactive()
)

use_notestar_makefile(path = ".")
}
\arguments{
\item{dir_project}{file-path to the base/root folder of the project. Defaults to
\code{"."} which is the current working directory.}

\item{dir_notebook}{Name of the directory containing the Rmd files. It should
be a relative path from the project root. Defaults to \code{"notebook"}}

\item{dir_md}{Name of the directory to contain md files (knitted Rmd files).
It should be a relative path from the project root. Defaults to
\code{"notebook/book"}.}

\item{notebook_helper}{Relative path to an R script to run before knitting
each Rmd file. Defaults to \code{"notebook/book/knitr-helpers.R"}}
}
\value{
\code{use_notestar_makefile()} creates a Makefile that will build or clean a
targets-based workflow.
}
\description{
Create a brand new notebook with default settings
}
